miyoo-mini: start importing from other experiments
This is a weird one, I've experimented with merging the partitions into
a single squashfs image for the rootfs, and it works quite well.

That other experiment is mostly a failure since it was meant to produce
a "FHS-exclusive" rootfs that stayed compatible with CFW (MiniUI, token
effort with OnionOS), though that bit fails a bit.

So I'm re-invigorating the effort **through celun**.