simulator/miyoo-mini-armv7l: Init
This target is one step closer to the real hardware.

But what good is this? Its main use case is to allow running *actual
armv7 binaries*, this way we can be reasonably confident that if up to
and including mounting the rootfs things work, everything else should
work on real hardware.