mkDotApp: proof of concept for app runtime self-containment
Note that this only does isolation of the *runtime*. What does this
mean? That we're only chrooting to make the root appear convenient for
our self-contained Nix-closure.

There are absolutely no security considerations here. Additionally, any
processes launched and not culled will keep running, which is most
likely undesirable for proper isolation.

Note again: this only is a way to conveniently ship a binary package,
and not at all a security layer.