miyoo/mini: Add dotAppToMiniUIPak
This packages a dotApp package into a form that is usable by MiniUI,
without any changes to MiniUI.

This includes a clumsy kernel module build (which will be broken out
into the general Miyoo Mini build later). This module is load-bearing as
it provides the devloop subsystem, which is missing from the vendor
kernel.

With this, we can mount the .app, chroot/run the entrypoint, and once
it's done we clean up behind ourselves.

Why? Two reasons. First, this could be used to ship games-os-built
packages for use on MiniUI. Second, and more important, this allows me
to test the user-facing apps on hardware without having to make it
unusable for playint with.

Though, really, the load-bearing part here is the loop kernel module.
With this, we can move forward with the plan to ship an alternative
rootfs that lives in a squashfs file on the SD card. This plan should
also be usable for other CFWs, even making it possible to ship a sort of
CFW selector at boot time.