stub: harden against slow storage
On some platforms (noticed on RK3566-based RG353V), it's possible that
the drivers for storage exhibits race condition with our init. Our init
does so little that it happens to try to mount before the block device
is there, but barely! It sometimes worked!

With this change, we wait 60 seconds, flailing around 10 times a second
hoping the block device is now there. This ought to be enough.