WIP: SDL: Use SDL1.2 with KMS/DRM
This uses the (rebased) OpenDingux fork, which adds KMS/DRM backend.
This, in turn, is required for use on the RG280M, as the fbcon backend
would require applications to scale themselves.

Though, KMS/DRM on other platforms may be useful, so let's aim to make
this the supported version.

WIP: This needs to use a published SDL fork.